Refactor classes and move to better packages.
